+++ src/main/java/xyz/pyrehaven/happyartillery/mixin/PlayerDropMixin.java
@@ -1,6 +1,7 @@
 package xyz.pyrehaven.happyartillery.mixin;
 
 import net.minecraft.server.level.ServerPlayer;
+import net.minecraft.util.Prediction;
 import net.minecraft.world.entity.item.ItemEntity;
 import net.minecraft.world.item.ItemStack;
 import org.spongepowered.asm.mixin.Mixin;
@@ -12,14 +13,14 @@ import xyz.pyrehaven.happyartillery.Controls;
 @Mixin(ServerPlayer.class)
 abstract class PlayerDropMixin {
     @Inject(
-            method = "drop(Lnet/minecraft/world/item/ItemStack;ZZ)"
+            method = "drop(Lnet/minecraft/world/item/ItemStack;ZLnet/minecraft/util/Prediction;)"
                     + "Lnet/minecraft/world/entity/item/ItemEntity;",
             at = @At("RETURN"),
             require = 1)
     private void happyArtillery$consumeMarkedDrop(
             ItemStack stack,
-            boolean randomThrow,
-            boolean retainOwnership,
+            boolean awardStats,
+            Prediction prediction,
             CallbackInfoReturnable<ItemEntity> callback) {
         Controls.consumeDroppedControl(stack, callback.getReturnValue());
     }